Prune redundant tests per handbook test-pruning guidelines
Same-entry-point duplicates removed: union-repair tests consolidated
into parse-raw-custom-tool-call.test.ts (the only file with the real-
object-passthrough case), loose-schema cases covered once by the quwin
repro in prompts-schema-handling, and the three-transport e2e reduced
to one wiring guard since the mapping itself is unit-tested next door.
