+++ packages/agent-runtime/src/__tests__/prompts-schema-handling.test.ts
@@ -646,105 +646,3 @@ describe('getToolSet: loose MCP schemas survive the point-of-use round-trip', ()
   })
 })
 
-// Some models hedge on union-typed parameters: when a schema says a param may
-// be a string OR an object (anyOf), the model sometimes emits the object as a
-// JSON-encoded string, because a string is unambiguously valid for the union.
-// The whole pipeline preserves that string faithfully, so the MCP server
-// receives a string where an object was meant - and since the union accepts
-// strings, nothing fails loudly. The tool-executor already repairs
-// double-encoded arguments at the top level; these tests pin the same repair
-// for nested, schema-guided cases. Found while digging into quwin's report
-// on issue #912.
-describe('parseRawCustomToolCall: schema-guided repair of string-encoded union members', () => {
-  const buildWithCustomTool = (inputSchema: unknown) => ({
-    customToolDefs: {
-      'loose-server__loose_union': {
-        description: 'Echoes back exactly the arguments it received.',
-        inputSchema: inputSchema as never,
-        endsAgentStep: false,
-      },
-    },
-    rawToolCall: {
-      toolName: 'loose-server__loose_union',
-      toolCallId: 'probe-1',
-    },
-  })
-
-  const unionSchema = {
-    $schema: 'http://json-schema.org/draft-07/schema#',
-    type: 'object',
-    properties: {
-      spec: {
-        anyOf: [{ type: 'string' }, { type: 'object', properties: { kind: { type: 'string' } }, additionalProperties: true }],
-        description: 'A string or an object. Either is accepted.',
-      },
-    },
-    required: ['spec'],
-    additionalProperties: true,
-  }
-
-  test('a JSON-encoded string for a union param with an object variant is decoded to an object', () => {
-    // Given a union schema (string | object) and a raw tool call whose
-    // union-typed parameter arrived as a JSON-encoded string,
-    // when the custom tool call is parsed,
-    // then the parameter is decoded to the object the model meant -
-    // matching what the same model emits for unambiguous object params.
-
-    // Arrange
-    const { customToolDefs, rawToolCall } = buildWithCustomTool(unionSchema)
-    const withInput = { ...rawToolCall, input: { spec: '{"kind": "unhinged-union-spec", "extra": 42}' } }
-
-    // Act
-    const result = parseRawCustomToolCall({ customToolDefs, rawToolCall: withInput })
-
-    // Assert
-    expect(result).toHaveProperty('input')
-    expect((result as { input: { spec: unknown } }).input.spec).toEqual({
-      kind: 'unhinged-union-spec',
-      extra: 42,
-    })
-  })
-
-  test('a non-JSON string for a union param stays a string', () => {
-    // Given the same union schema and a parameter that is a plain string
-    // (not JSON-encoded),
-    // when parsed,
-    // then the string is preserved - the string branch of the union is a
-    // legitimate choice and must not be mangled.
-
-    // Arrange
-    const { customToolDefs, rawToolCall } = buildWithCustomTool(unionSchema)
-    const withInput = { ...rawToolCall, input: { spec: 'plain-string-variant' } }
-
-    // Act
-    const result = parseRawCustomToolCall({ customToolDefs, rawToolCall: withInput })
-
-    // Assert
-    expect((result as { input: { spec: unknown } }).input.spec).toBe('plain-string-variant')
-  })
-
-  test('a JSON-encoded string for a plain string-typed param is NOT decoded', () => {
-    // Given a schema whose param is a plain string (no object variant) and a
-    // value that happens to be JSON-encoded,
-    // when parsed,
-    // then the string stays a string - the repair must be guided by the
-    // schema, or tools whose string params legitimately contain JSON (like
-    // evaluate_script source) would be corrupted.
-
-    // Arrange
-    const stringOnlySchema = {
-      type: 'object',
-      properties: { code: { type: 'string' } },
-      required: ['code'],
-      additionalProperties: false,
-    }
-    const { customToolDefs, rawToolCall } = buildWithCustomTool(stringOnlySchema)
-    const withInput = { ...rawToolCall, input: { code: '{"looks": "like json"}' } }
-
-    // Act
-    const result = parseRawCustomToolCall({ customToolDefs, rawToolCall: withInput })
-
-    // Assert
-    expect((result as { input: { code: unknown } }).input.code).toBe('{"looks": "like json"}')
-  })
-})
